+++ app/src/main/java/com/example/turystycznezaglebie/Greedy.java
@@ -55,7 +55,7 @@ public class Greedy extends Algorithm{
     }
 
     @Override
-    public int findWayMultimodal(int startPoint0, int timeMax, long calculation_time) {
+    public float findWayMultimodal(int startPoint0, int timeMax, long calculation_time) {
         return 0;
     }
 }
